// src/interfaces.ts
export interface ICulversLocation {
  dineInHours: string;
  driveThruHours: string;
  curbsideHours: string;
  lateNightDriveThruHours: string;
  onlineOrderStatus: number;
  flavorOfDayName: string;
  flavorOfTheDayDescription: string;
  flavorOfDaySlug: string;
  openDate: string;
  isTemporarilyClosed: boolean;
  utcOffset: number;
  street: string;
  state: string;
  city: string;
  postalCode: string;
  oloId: string;
  slug: string;
  jobsearchurl: string;
  handoffOptions: string;
  restKeyId: string;
  restaurantNumber: string;
}